Virtual threads: integration tests refactoring
- move the logic from AssertHelpers to a shared component - VirtualThreadsAssertions located in quarkus-test-vertx
