+++ integration-tests/virtual-threads/amqp-virtual-threads/src/main/java/io/quarkus/it/vthreads/amqp/PriceConsumer.java
@@ -1,8 +1,5 @@
 package io.quarkus.it.vthreads.amqp;
 
-import static io.quarkus.it.vthreads.amqp.AssertHelper.assertThatItRunsOnADuplicatedContext;
-import static io.quarkus.it.vthreads.amqp.AssertHelper.assertThatItRunsOnVirtualThread;
-
 import java.util.Random;
 import java.util.concurrent.CompletionStage;
 import java.util.concurrent.atomic.AtomicInteger;
@@ -14,6 +11,7 @@ import org.eclipse.microprofile.reactive.messaging.Message;
 import org.eclipse.microprofile.reactive.messaging.Outgoing;
 import org.eclipse.microprofile.rest.client.inject.RestClient;
 
+import io.quarkus.test.vertx.VirtualThreadsAssertions;
 import io.smallrye.common.annotation.RunOnVirtualThread;
 
 @ApplicationScoped
@@ -25,12 +23,12 @@ public class PriceConsumer {
     @Incoming("prices")
     @RunOnVirtualThread
     public CompletionStage<Void> consume(Message<Double> msg) {
-        assertThatItRunsOnVirtualThread();
-        assertThatItRunsOnADuplicatedContext();
+        VirtualThreadsAssertions.assertThatItRunsOnVirtualThread();
+        VirtualThreadsAssertions.assertThatItRunsOnADuplicatedContext();
         double price = msg.getPayload();
         alertService.alertMessage(price);
         return msg.ack().thenAccept(x -> {
-            assertThatItRunsOnADuplicatedContext();
+            VirtualThreadsAssertions.assertThatItRunsOnADuplicatedContext();
             // While the ack always runs on event loop thread
             // the post-ack may run on the processing virtual-thread which executed the method.
         });
@@ -39,8 +37,8 @@ public class PriceConsumer {
     @Incoming("prices")
     @RunOnVirtualThread
     public void consume(double price) {
-        assertThatItRunsOnVirtualThread();
-        assertThatItRunsOnADuplicatedContext();
+        VirtualThreadsAssertions.assertThatItRunsOnVirtualThread();
+        VirtualThreadsAssertions.assertThatItRunsOnADuplicatedContext();
         alertService.alert(price);
     }
 
@@ -50,7 +48,7 @@ public class PriceConsumer {
     @Outgoing("prices-out")
     @RunOnVirtualThread
     public Message<Double> randomPriceGenerator() {
-        assertThatItRunsOnVirtualThread();
+        VirtualThreadsAssertions.assertThatItRunsOnVirtualThread();
         return Message.of(r.nextDouble() * 10 * i.incrementAndGet());
     }
