+++ integration-tests/virtual-threads/jms-virtual-threads/src/main/java/io/quarkus/it/vthreads/jms/PriceConsumer.java
@@ -1,7 +1,5 @@
 package io.quarkus.it.vthreads.jms;
 
-import static io.quarkus.it.vthreads.jms.AssertHelper.assertThatItRunsOnVirtualThread;
-
 import java.util.Random;
 import java.util.concurrent.CompletionStage;
 import java.util.concurrent.atomic.AtomicInteger;
@@ -13,6 +11,7 @@ import org.eclipse.microprofile.reactive.messaging.Message;
 import org.eclipse.microprofile.reactive.messaging.Outgoing;
 import org.eclipse.microprofile.rest.client.inject.RestClient;
 
+import io.quarkus.test.vertx.VirtualThreadsAssertions;
 import io.smallrye.common.annotation.RunOnVirtualThread;
 
 @ApplicationScoped
@@ -24,7 +23,7 @@ public class PriceConsumer {
     @Incoming("prices")
     @RunOnVirtualThread
     public CompletionStage<Void> consume(Message<Double> msg) {
-        assertThatItRunsOnVirtualThread();
+        VirtualThreadsAssertions.assertThatItRunsOnVirtualThread();
         double price = msg.getPayload();
         alertService.alertMessage(price);
         return msg.ack();
@@ -33,7 +32,7 @@ public class PriceConsumer {
     @Incoming("prices")
     @RunOnVirtualThread
     public void consume(double price) {
-        assertThatItRunsOnVirtualThread();
+        VirtualThreadsAssertions.assertThatItRunsOnVirtualThread();
         alertService.alert(price);
     }
 
@@ -43,7 +42,7 @@ public class PriceConsumer {
     @Outgoing("prices-out")
     @RunOnVirtualThread
     public Message<Double> randomPriceGenerator() {
-        assertThatItRunsOnVirtualThread();
+        VirtualThreadsAssertions.assertThatItRunsOnVirtualThread();
         return Message.of(r.nextDouble() * 10 * i.incrementAndGet());
     }
