+++ integration-tests/virtual-threads/reactive-routes-virtual-threads/src/test/java/io/quarkus/virtual/vertx/web/RunOnVirtualThreadTest.java
@@ -1,6 +1,7 @@
 package io.quarkus.virtual.vertx.web;
 
 import static io.restassured.RestAssured.get;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
 
 import org.junit.jupiter.api.Test;
@@ -15,11 +16,20 @@ import io.quarkus.test.junit5.virtual.VirtualThreadUnit;
 class RunOnVirtualThreadTest {
 
     @Test
-    void testRoute() {
+    void testRouteOnVirtualThread() {
         String bodyStr = get("/hello").then().statusCode(200).extract().asString();
         // Each VT has a unique name in quarkus
         assertNotEquals(bodyStr, get("/hello").then().statusCode(200).extract().asString());
+    }
 
+    @Test
+    void testRouteOnEventLoop() {
+        assertEquals("pong", get("/ping").then().statusCode(200).extract().asString());
+    }
+
+    @Test
+    void testRouteOnWorker() {
+        assertEquals("pong", get("/blocking-ping").then().statusCode(200).extract().asString());
     }
 
 }
