+++ integration-tests/virtual-threads/redis-virtual-threads/src/main/java/io/quarkus/virtual/redis/RedisResource.java
@@ -8,6 +8,7 @@ import jakarta.ws.rs.Path;
 import io.quarkus.cache.CacheResult;
 import io.quarkus.redis.datasource.RedisDataSource;
 import io.quarkus.redis.datasource.hash.HashCommands;
+import io.quarkus.test.vertx.VirtualThreadsAssertions;
 import io.smallrye.common.annotation.RunOnVirtualThread;
 
 @Path("/")
@@ -22,7 +23,7 @@ public class RedisResource {
 
     @GET
     public String testRedis() {
-        AssertHelper.assertEverything();
+        VirtualThreadsAssertions.assertEverything();
         String value = UUID.randomUUID().toString();
         hash.hset("test", "test", value);
