+++ integration-tests/virtual-threads/resteasy-reactive-virtual-threads/src/main/java/io/quarkus/virtual/rr/Filters.java
@@ -8,15 +8,14 @@ import org.jboss.logmanager.MDC;
 import org.jboss.resteasy.reactive.server.ServerRequestFilter;
 import org.jboss.resteasy.reactive.server.ServerResponseFilter;
 
+import io.quarkus.test.vertx.VirtualThreadsAssertions;
 import io.vertx.core.Vertx;
 
 public class Filters {
     @ServerRequestFilter(nonBlocking = true)
     public void request(ContainerRequestContext requestContext) {
         if (requestContext.getUriInfo().getPath().contains("/filter")) {
-            AssertHelper.assertNotOnVirtualThread();
-            AssertHelper.assertThatItRunsOnADuplicatedContext();
-            AssertHelper.assertThatTheRequestScopeIsActive();
+            VirtualThreadsAssertions.assertWorkerOrEventLoopThread();
             MDC.put("mdc", "test");
             CDI.current().select(Counter.class).get().increment();
             Vertx.currentContext().putLocal("filter", "test");
@@ -26,7 +25,7 @@ public class Filters {
     @ServerResponseFilter
     public void getFilter(ContainerResponseContext responseContext) {
         if (responseContext.getHeaders().get("X-filter") != null) {
-            AssertHelper.assertEverything();
+            VirtualThreadsAssertions.assertEverything();
             // the request filter, the method, and here.
             assert CDI.current().select(Counter.class).get().increment() == 3;
             assert Vertx.currentContext().getLocal("test").equals("test test");
