+++ integration-tests/virtual-threads/vertx-event-bus-virtual-threads/src/main/java/io/quarkus/virtual/vertx/EventBusConsumer.java
@@ -5,6 +5,7 @@ import java.util.concurrent.CopyOnWriteArrayList;
 
 import jakarta.enterprise.context.ApplicationScoped;
 
+import io.quarkus.test.vertx.VirtualThreadsAssertions;
 import io.quarkus.vertx.ConsumeEvent;
 import io.smallrye.common.annotation.RunOnVirtualThread;
 
@@ -16,14 +17,14 @@ public class EventBusConsumer {
     @ConsumeEvent("one-way")
     @RunOnVirtualThread
     void receive(String m) {
-        AssertHelper.assertEverything();
+        VirtualThreadsAssertions.assertEverything();
         ONE_WAY.add(m);
     }
 
     @ConsumeEvent("request-reply")
     @RunOnVirtualThread
     String process(String m) {
-        AssertHelper.assertEverything();
+        VirtualThreadsAssertions.assertEverything();
         return m.toUpperCase();
     }
