+++ integration-tests/virtual-threads/virtual-threads-disabled/src/main/java/io/quarkus/virtual/disabled/MyResource.java
@@ -4,6 +4,7 @@ import jakarta.ws.rs.GET;
 import jakarta.ws.rs.POST;
 import jakarta.ws.rs.Path;
 
+import io.quarkus.test.vertx.VirtualThreadsAssertions;
 import io.smallrye.common.annotation.NonBlocking;
 import io.smallrye.common.annotation.RunOnVirtualThread;
 import io.smallrye.mutiny.Multi;
@@ -21,14 +22,14 @@ public class MyResource {
     @GET
     @RunOnVirtualThread
     public String testGet() {
-        AssertHelper.assertWorkerOrEventLoopThread();
+        VirtualThreadsAssertions.assertWorkerOrEventLoopThread();
         return "hello-" + counter.increment();
     }
 
     @POST
     @RunOnVirtualThread
     public String testPost(String body) {
-        AssertHelper.assertWorkerOrEventLoopThread();
+        VirtualThreadsAssertions.assertWorkerOrEventLoopThread();
         return body + "-" + counter.increment();
     }
 
@@ -36,34 +37,28 @@ public class MyResource {
     @NonBlocking
     @Path("/non-blocking")
     public String testNonBlocking() {
-        AssertHelper.assertWorkerOrEventLoopThread();
+        VirtualThreadsAssertions.assertWorkerOrEventLoopThread();
         return "ok";
     }
 
     @GET
     @Path("/uni")
     public Uni<String> testUni() {
-        AssertHelper.assertThatTheRequestScopeIsActive();
-        AssertHelper.assertThatItRunsOnADuplicatedContext();
-        AssertHelper.assertNotOnVirtualThread();
+        VirtualThreadsAssertions.assertWorkerOrEventLoopThread();
         return Uni.createFrom().item("ok");
     }
 
     @GET
     @Path("/multi")
     public Multi<String> testMulti() {
-        AssertHelper.assertThatTheRequestScopeIsActive();
-        AssertHelper.assertThatItRunsOnADuplicatedContext();
-        AssertHelper.assertNotOnVirtualThread();
+        VirtualThreadsAssertions.assertWorkerOrEventLoopThread();
         return Multi.createFrom().items("o", "k");
     }
 
     @GET
     @Path("/blocking")
     public String testBlocking() {
-        AssertHelper.assertThatTheRequestScopeIsActive();
-        AssertHelper.assertThatItRunsOnADuplicatedContext();
-        AssertHelper.assertNotOnVirtualThread();
+        VirtualThreadsAssertions.assertWorkerOrEventLoopThread();
         return "hello-" + counter.increment();
     }
